requirements.txt
Flask==0.12
gunicorn==19.6.0
requests==2.25.0
psycopg2-binary==2.8.6
dj-database-url==0.5.0
boto3==1.16.32
pandas=1.1.4
matplotlib=3.3.3